Revert "Ifpack2: Add Schur complements to the BlockTriDiag solver. (#12344)"
This reverts commit 6ad4a253d7904f7e0d5d10cfe5a457fe199114bb.
