+++ packages/ifpack2/src/Ifpack2_BlockTriDiContainer_decl.hpp
@@ -228,7 +228,6 @@ namespace Ifpack2 {
     ///   purposes only.
     BlockTriDiContainer (const Teuchos::RCP<const row_matrix_type>& matrix,
                          const Teuchos::Array<Teuchos::Array<local_ordinal_type> >& partitions,
-                         const int n_subparts_per_part = 1,
                          bool overlapCommAndComp = false, bool useSequentialMethod = false);
 
     //! Destructor (declared virtual for memory safety of derived classes).
@@ -390,13 +389,11 @@ namespace Ifpack2 {
 
     // hide details of impl using ImplObj; finally I understand why AMB did that way.
     Teuchos::RCP<BlockTriDiContainerDetails::ImplObject<MatrixType> > impl_;
-    int n_subparts_per_part_;
     
     // initialize distributed and local objects
     void initInternal (const Teuchos::RCP<const row_matrix_type>& matrix,
                        const Teuchos::Array<Teuchos::Array<local_ordinal_type> >& partitions,
                        const Teuchos::RCP<const import_type> &importer,
-                       const int n_subparts_per_part,
                        const bool overlapCommAndComp,
                        const bool useSeqMethod);
