+++ packages/ifpack2/src/Ifpack2_BlockTriDiContainer_def.hpp
@@ -81,50 +81,34 @@ namespace Ifpack2 {
   ::initInternal (const Teuchos::RCP<const row_matrix_type>& matrix,
                   const Teuchos::Array<Teuchos::Array<local_ordinal_type> >& partitions,
                   const Teuchos::RCP<const import_type>& importer,
-                  const int n_subparts_per_part,
                   const bool overlapCommAndComp,
                   const bool useSeqMethod) 
   {
-    IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::initInternal");
-    n_subparts_per_part_ = n_subparts_per_part;
-
     // create pointer of impl
-    {
-      IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::createImpl");
-      impl_ = Teuchos::rcp(new BlockTriDiContainerDetails::ImplObject<MatrixType>());
-      IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
-    }
+    impl_ = Teuchos::rcp(new BlockTriDiContainerDetails::ImplObject<MatrixType>());
 
-    using impl_type = BlockHelperDetails::ImplType<MatrixType>;
+    using impl_type = BlockTriDiContainerDetails::ImplType<MatrixType>;
     // using block_crs_matrix_type = typename impl_type::tpetra_block_crs_matrix_type;
 
-    {
-      IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::setA");
-      impl_->A = Teuchos::rcp_dynamic_cast<const block_crs_matrix_type>(matrix);
-      TEUCHOS_TEST_FOR_EXCEPT_MSG
-        (impl_->A.is_null(), "BlockTriDiContainer currently supports Tpetra::BlockCrsMatrix only.");
-      IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
-    }
+    impl_->A = Teuchos::rcp_dynamic_cast<const block_crs_matrix_type>(matrix);
+    TEUCHOS_TEST_FOR_EXCEPT_MSG
+      (impl_->A.is_null(), "BlockTriDiContainer currently supports Tpetra::BlockCrsMatrix only.");
 
     impl_->tpetra_importer = Teuchos::null;
     impl_->async_importer  = Teuchos::null;
     
     if (useSeqMethod)
     {
-      IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::createBlockCrsTpetraImporter useSeqMethod");
       if (importer.is_null()) // there is no given importer, then create one
         impl_->tpetra_importer = BlockTriDiContainerDetails::createBlockCrsTpetraImporter<MatrixType>(impl_->A);
       else
         impl_->tpetra_importer = importer; // if there is a given importer, use it
-      IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
     }
     else
     {
-      IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::createBlockCrsTpetraImporter");
       //Leave tpetra_importer null even if user provided an importer.
       //It is not used in the performant codepath (!useSeqMethod)
       impl_->async_importer = BlockTriDiContainerDetails::createBlockCrsAsyncImporter<MatrixType>(impl_->A);
-      IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
     }
 
     // as a result, there are 
@@ -134,26 +118,13 @@ namespace Ifpack2 {
 
     // temporary disabling 
     impl_->overlap_communication_and_computation = overlapCommAndComp;
+    
+    impl_->Z = typename impl_type::tpetra_multivector_type();
+    impl_->W = typename impl_type::impl_scalar_type_1d_view();
 
-    {
-      IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::createZ");
-      impl_->Z = typename impl_type::tpetra_multivector_type();
-      IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
-    }
-    {
-      IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::createW");
-      impl_->W = typename impl_type::impl_scalar_type_1d_view();
-      IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
-    }
-
-    {
-      IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::createPartInterfaceBlockTridiagsNormManager");
-      impl_->part_interface  = BlockTriDiContainerDetails::createPartInterface<MatrixType>(impl_->A, partitions, n_subparts_per_part_);
-      impl_->block_tridiags  = BlockTriDiContainerDetails::createBlockTridiags<MatrixType>(impl_->part_interface);
-      impl_->norm_manager    = BlockHelperDetails::NormManager<MatrixType>(impl_->A->getComm());
-      IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
-    }
-    IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
+    impl_->part_interface  = BlockTriDiContainerDetails::createPartInterface<MatrixType>(impl_->A, partitions);
+    impl_->block_tridiags  = BlockTriDiContainerDetails::createBlockTridiags<MatrixType>(impl_->part_interface);
+    impl_->norm_manager    = BlockTriDiContainerDetails::NormManager<MatrixType>(impl_->A->getComm());
   }
 
   template <typename MatrixType>
@@ -161,12 +132,11 @@ namespace Ifpack2 {
   BlockTriDiContainer<MatrixType, BlockTriDiContainerDetails::ImplSimdTag>
   ::clearInternal ()
   {
-    IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::clearInternal");
-    using impl_type = BlockHelperDetails::ImplType<MatrixType>;
-    using part_interface_type = BlockHelperDetails::PartInterface<MatrixType>;
+    using impl_type = BlockTriDiContainerDetails::ImplType<MatrixType>;
+    using part_interface_type = BlockTriDiContainerDetails::PartInterface<MatrixType>;
     using block_tridiags_type = BlockTriDiContainerDetails::BlockTridiags<MatrixType>;
-    using amd_type = BlockHelperDetails::AmD<MatrixType>;
-    using norm_manager_type = BlockHelperDetails::NormManager<MatrixType>;
+    using amd_type = BlockTriDiContainerDetails::AmD<MatrixType>;
+    using norm_manager_type = BlockTriDiContainerDetails::NormManager<MatrixType>;
     
     impl_->A = Teuchos::null;
     impl_->tpetra_importer = Teuchos::null;
@@ -182,7 +152,6 @@ namespace Ifpack2 {
     impl_->norm_manager    = norm_manager_type();
 
     impl_ = Teuchos::null;
-    IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
   }
 
   template <typename MatrixType>
@@ -193,25 +162,19 @@ namespace Ifpack2 {
                        bool pointIndexed)
     : Container<MatrixType>(matrix, partitions, pointIndexed)
   {
-    IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::BlockTriDiContainer");
     const bool useSeqMethod = false;
     const bool overlapCommAndComp = false;
-    initInternal(matrix, partitions, importer, 1, overlapCommAndComp, useSeqMethod);
-    IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
+    initInternal(matrix, partitions, importer, overlapCommAndComp, useSeqMethod);
   }
 
   template <typename MatrixType>
   BlockTriDiContainer<MatrixType, BlockTriDiContainerDetails::ImplSimdTag>
   ::BlockTriDiContainer (const Teuchos::RCP<const row_matrix_type>& matrix,
                        const Teuchos::Array<Teuchos::Array<local_ordinal_type> >& partitions,
-                       const int n_subparts_per_part,
-                       const bool overlapCommAndComp, 
-                       const bool useSeqMethod)
+                       const bool overlapCommAndComp, const bool useSeqMethod)
     : Container<MatrixType>(matrix, partitions, false)
   {
-    IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::BlockTriDiContainer");
-    initInternal(matrix, partitions, Teuchos::null, n_subparts_per_part, overlapCommAndComp, useSeqMethod);
-    IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
+    initInternal(matrix, partitions, Teuchos::null, overlapCommAndComp, useSeqMethod);
   }
 
   template <typename MatrixType>
@@ -233,7 +196,6 @@ namespace Ifpack2 {
   BlockTriDiContainer<MatrixType, BlockTriDiContainerDetails::ImplSimdTag>
   ::initialize ()
   {
-    IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::initialize");
     this->IsInitialized_ = true;
     // We assume that if you called this method, you intend to recompute
     // everything.
@@ -246,7 +208,6 @@ namespace Ifpack2 {
          impl_->a_minus_d, 
          impl_->overlap_communication_and_computation);    
     }
-    IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
   }
 
   template <typename MatrixType>
@@ -254,7 +215,6 @@ namespace Ifpack2 {
   BlockTriDiContainer<MatrixType, BlockTriDiContainerDetails::ImplSimdTag>
   ::compute ()
   {
-    IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::compute");
     this->IsComputed_ = false;
     if (!this->isInitialized())
       this->initialize();
@@ -265,7 +225,6 @@ namespace Ifpack2 {
          Kokkos::ArithTraits<magnitude_type>::zero());
     }
     this->IsComputed_ = true;
-    IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
   }
 
   template <typename MatrixType>
@@ -273,12 +232,10 @@ namespace Ifpack2 {
   BlockTriDiContainer<MatrixType, BlockTriDiContainerDetails::ImplSimdTag>
   ::clearBlocks ()
   {
-    IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::clearBlocks");
     clearInternal();
     this->IsInitialized_ = false;
     this->IsComputed_ = false;
     Container<MatrixType>::clearBlocks();
-    IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
   }
 
   template <typename MatrixType>
@@ -287,7 +244,6 @@ namespace Ifpack2 {
   ::applyInverseJacobi (const mv_type& X, mv_type& Y, scalar_type dampingFactor,
                         bool zeroStartingSolution, int numSweeps) const
   {
-    IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::applyInverseJacobi");
     const magnitude_type tol = Kokkos::ArithTraits<magnitude_type>::zero();
     const int check_tol_every = 1;
 
@@ -305,7 +261,6 @@ namespace Ifpack2 {
        numSweeps,
        tol,
        check_tol_every);
-    IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
   }
 
   template <typename MatrixType>
@@ -321,7 +276,6 @@ namespace Ifpack2 {
   BlockTriDiContainer<MatrixType, BlockTriDiContainerDetails::ImplSimdTag>
   ::compute (const ComputeParameters& in)
   {
-    IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::compute");
     this->IsComputed_ = false;
     if (!this->isInitialized())
       this->initialize();
@@ -332,7 +286,6 @@ namespace Ifpack2 {
          in.addRadiallyToDiagonal);
     }
     this->IsComputed_ = true;
-    IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
   }
 
   template <typename MatrixType>
@@ -351,7 +304,6 @@ namespace Ifpack2 {
   ::applyInverseJacobi (const mv_type& X, mv_type& Y, 
                         const ApplyParameters& in) const
   {
-    IFPACK2_BLOCKHELPER_TIMER("BlockTriDiContainer::applyInverseJacobi");
     int r_val = 0;
     {
       r_val = BlockTriDiContainerDetails::applyInverseJacobi<MatrixType>
@@ -369,7 +321,6 @@ namespace Ifpack2 {
          in.tolerance,
          in.checkToleranceEvery);
     }
-    IFPACK2_BLOCKHELPER_TIMER_FENCE(typename BlockHelperDetails::ImplType<MatrixType>::execution_space)
     return r_val;
   }
