+++ packages/ifpack2/test/unit_tests/Ifpack2_UnitTestBlockTriDiContainerUtil.hpp
@@ -186,7 +186,7 @@ struct BlockTriDiContainerTester {
     Teuchos::Array<Teuchos::Array<LO> > parts;
     make_parts(sb, sbp, *A, nonuniform_lines, jacobi, parts);
     return Teuchos::rcp(new Ifpack2::BlockTriDiContainer<Tpetra_RowMatrix>(
-                          A, parts, 1, overlap_comm, seq_method));
+                          A, parts, overlap_comm, seq_method));
   }
 
   static Int
